refactor: split social raw migration modules
