+++ .agents/scripts/knowledge_social_store.py
@@ -9,14 +9,22 @@ import gzip
 import hashlib
 import json
 import os
-import re
 import sqlite3
 from pathlib import Path
 
+from _knowledge_social_store_migration import (
+    migrate_fetch_batches_v6 as _migrate_fetch_batches_v6,
+    recover_orphaned_fetch_batches_v6 as _recover_orphaned_fetch_batches_v6,
+)
 from _knowledge_social_store_schema import (
     add_outbound_v4_columns as _add_outbound_v4_columns,
     add_source_v5_columns as _add_source_v5_columns,
     add_sync_run_v2_columns as _add_sync_run_v2_columns,
+)
+from _knowledge_social_store_support import (
+    MAX_RAW_BATCH_BYTES,
+    OPAQUE_ID,
+    SocialStoreError,
 )
 from knowledge_corpus_context import (
     CatalogError,
@@ -26,27 +34,7 @@ from knowledge_corpus_context import (
 
 SCHEMA_VERSION = 6
 SCHEMA_VERSION_SQL = "PRAGMA user_version=6"
-OPAQUE_ID = re.compile(r"^[A-Za-z0-9][A-Za-z0-9_-]{2,127}$")
-SHA256_HEX = re.compile(r"^[0-9a-f]{64}$")
 SQLITE_MUTABLE_SIDECARS = ("-journal", "-shm", "-wal")
-MAX_RAW_BATCH_BYTES = 64 * 1024 * 1024
-COLLECTOR_ENVELOPE_FIELDS = frozenset(
-    {
-        "provider",
-        "connection_id",
-        "stream",
-        "observed_at",
-        "request_hash",
-        "response_sha256",
-        "response",
-    }
-)
-INVALID_RAW_PATH = "legacy social raw evidence path is invalid"
-INVALID_RAW_METADATA = "legacy social raw evidence metadata is invalid"
-
-
-class SocialStoreError(RuntimeError):
-    """Raised when private social storage cannot be used safely."""
 
 
 def private_directory(root: Path, relative: Path) -> Path:
@@ -277,341 +265,6 @@ def create_fts(connection: sqlite3.Connection) -> None:
         )
     except sqlite3.OperationalError as error:
         raise SocialStoreError("SQLite runtime does not provide required FTS5") from error
-
-
-def _store_root(connection: sqlite3.Connection) -> Path:
-    database_rows = connection.execute("PRAGMA database_list").fetchall()
-    database_file = next(
-        (str(row["file"]) for row in database_rows if row["name"] == "main"), ""
-    )
-    if not database_file:
-        raise SocialStoreError("legacy social data requires a file-backed store")
-    return Path(database_file).resolve(strict=True).parent.parent
-
-
-def _validated_raw_ref(blob_ref: str) -> tuple[Path, str, str, str]:
-    relative = Path(blob_ref)
-    if relative.is_absolute():
-        raise SocialStoreError(INVALID_RAW_PATH)
-    if len(relative.parts) != 6:
-        raise SocialStoreError(INVALID_RAW_PATH)
-    prefix, provider, connection_id, filename = (
-        relative.parts[:3],
-        relative.parts[3],
-        relative.parts[4],
-        relative.parts[5],
-    )
-    if prefix != ("sources", "social", "raw"):
-        raise SocialStoreError(INVALID_RAW_PATH)
-    if OPAQUE_ID.fullmatch(provider) is None:
-        raise SocialStoreError(INVALID_RAW_PATH)
-    if OPAQUE_ID.fullmatch(connection_id) is None:
-        raise SocialStoreError(INVALID_RAW_PATH)
-    if re.fullmatch(r"[0-9a-f]{64}\.json\.gz", filename) is None:
-        raise SocialStoreError(INVALID_RAW_PATH)
-    return relative, provider, connection_id, filename
-
-
-def _read_private_raw_file(root: Path, relative: Path) -> bytes:
-    raw_root = root / "sources" / "social" / "raw"
-    path = root / relative
-    try:
-        resolved = path.resolve(strict=True)
-        if resolved != Path(os.path.abspath(path)):
-            raise SocialStoreError("legacy social raw evidence contains a symlink")
-        resolved.relative_to(raw_root.resolve(strict=True))
-        validate_directory(path.parent.parent, "social provider directory", repair=False)
-        validate_directory(path.parent, "social connection directory", repair=False)
-        validate_private_file(path, "legacy social raw evidence", repair=False)
-        with gzip.open(path, "rb") as source:
-            payload = source.read(MAX_RAW_BATCH_BYTES + 1)
-    except (CatalogError, OSError, EOFError, ValueError) as error:
-        raise SocialStoreError("legacy social raw evidence is unsafe") from error
-    if len(payload) > MAX_RAW_BATCH_BYTES:
-        raise SocialStoreError("legacy social raw evidence exceeds the size limit")
-    return payload
-
-
-def _read_raw_payload(
-    connection: sqlite3.Connection, blob_ref: str
-) -> tuple[bytes, str, str, str]:
-    relative, provider, connection_id, filename = _validated_raw_ref(blob_ref)
-    payload = _read_private_raw_file(_store_root(connection), relative)
-    digest = hashlib.sha256(payload).hexdigest()
-    if filename != f"{digest}.json.gz":
-        raise SocialStoreError("legacy social raw evidence hash does not match")
-    return payload, digest, provider, connection_id
-
-
-def _decoded_collector_envelope(payload: bytes) -> dict[str, object] | None:
-    try:
-        envelope = json.loads(payload)
-    except (UnicodeDecodeError, json.JSONDecodeError):
-        return None
-    if not isinstance(envelope, dict):
-        return None
-    if set(envelope) != COLLECTOR_ENVELOPE_FIELDS:
-        return None
-    return envelope
-
-
-def _validate_collector_scope(
-    envelope: dict[str, object], payload: bytes, provider: str, connection_id: str
-) -> None:
-    canonical = json.dumps(
-        envelope, ensure_ascii=False, sort_keys=True, separators=(",", ":")
-    ).encode("utf-8")
-    if canonical != payload:
-        raise SocialStoreError(INVALID_RAW_METADATA)
-    if envelope["provider"] != provider:
-        raise SocialStoreError(INVALID_RAW_METADATA)
-    if envelope["connection_id"] != connection_id:
-        raise SocialStoreError(INVALID_RAW_METADATA)
-
-
-def _validate_collector_fields(envelope: dict[str, object]) -> None:
-    for field in ("stream", "observed_at"):
-        value = envelope[field]
-        if not isinstance(value, str) or not value:
-            raise SocialStoreError(INVALID_RAW_METADATA)
-    for field in ("request_hash", "response_sha256"):
-        value = envelope[field]
-        if not isinstance(value, str) or SHA256_HEX.fullmatch(value) is None:
-            raise SocialStoreError(INVALID_RAW_METADATA)
-
-
-def _validate_collector_response(envelope: dict[str, object]) -> None:
-    response_hash = str(envelope["response_sha256"])
-    response = json.dumps(
-        envelope["response"], ensure_ascii=False, sort_keys=True, separators=(",", ":")
-    ).encode("utf-8")
-    if hashlib.sha256(response).hexdigest() != response_hash:
-        raise SocialStoreError("legacy social response hash does not match")
-
-
-def _validate_collector_envelope(
-    envelope: dict[str, object], payload: bytes, provider: str, connection_id: str
-) -> None:
-    _validate_collector_scope(envelope, payload, provider, connection_id)
-    _validate_collector_fields(envelope)
-    _validate_collector_response(envelope)
-
-
-def _collector_envelope(
-    payload: bytes, provider: str, connection_id: str, *, required: bool
-) -> dict[str, object] | None:
-    envelope = _decoded_collector_envelope(payload)
-    if envelope is None:
-        if not required:
-            return None
-        raise SocialStoreError(INVALID_RAW_METADATA)
-    _validate_collector_envelope(envelope, payload, provider, connection_id)
-    return envelope
-
-
-def _canonical_fetch_identity_v6(
-    connection: sqlite3.Connection, row: sqlite3.Row
-) -> tuple[str, str]:
-    payload, digest, provider, connection_id = _read_raw_payload(
-        connection, str(row["blob_ref"])
-    )
-    envelope = _collector_envelope(
-        payload, provider, connection_id, required=False
-    )
-    if envelope is None:
-        if str(row["batch_id"]) != digest:
-            raise SocialStoreError("legacy fetch batch raw identity cannot be migrated")
-        return digest, str(row["response_hash"])
-    stored_scope = (
-        row["provider"],
-        row["connection_id"],
-        row["stream"],
-        row["request_hash"],
-        row["completed_at"],
-    )
-    raw_scope = (
-        provider,
-        connection_id,
-        envelope["stream"],
-        envelope["request_hash"],
-        envelope["observed_at"],
-    )
-    if stored_scope != raw_scope:
-        raise SocialStoreError("legacy fetch batch metadata conflicts with raw evidence")
-    return digest, str(envelope["response_sha256"])
-
-
-def _canonical_fetch_rows_v6(
-    connection: sqlite3.Connection,
-) -> list[tuple[sqlite3.Row, str, str]]:
-    rows = connection.execute("SELECT * FROM fetch_batches ORDER BY batch_id").fetchall()
-    return [
-        (row, *_canonical_fetch_identity_v6(connection, row)) for row in rows
-    ]
-
-
-def _deduplicate_fetch_rows_v6(
-    migrated: list[tuple[sqlite3.Row, str, str]],
-) -> dict[str, tuple[sqlite3.Row, str]]:
-    canonical: dict[str, tuple[sqlite3.Row, str]] = {}
-    signatures: dict[str, tuple[object, ...]] = {}
-    for row, batch_id, response_hash in migrated:
-        metadata = (
-            row["provider"],
-            row["connection_id"],
-            row["stream"],
-            row["request_hash"],
-            row["blob_ref"],
-            row["resource_count"],
-            row["budget_units"],
-            row["started_at"],
-            row["completed_at"],
-            row["terminal_status"],
-            response_hash,
-        )
-        existing = signatures.setdefault(batch_id, metadata)
-        if existing != metadata:
-            raise SocialStoreError("legacy fetch batch aliases have conflicting metadata")
-        canonical.setdefault(batch_id, (row, response_hash))
-    return canonical
-
-
-def _create_fetch_batches_v6(
-    connection: sqlite3.Connection,
-    canonical: dict[str, tuple[sqlite3.Row, str]],
-) -> None:
-    connection.execute(
-        """CREATE TABLE fetch_batches_v6 (
-             batch_id TEXT PRIMARY KEY, provider TEXT NOT NULL,
-             connection_id TEXT NOT NULL, stream TEXT NOT NULL,
-             request_hash TEXT, response_hash TEXT NOT NULL, blob_ref TEXT NOT NULL,
-             resource_count INTEGER NOT NULL, budget_units INTEGER NOT NULL DEFAULT 0,
-             started_at TEXT, completed_at TEXT NOT NULL, terminal_status TEXT NOT NULL,
-             evidence_id TEXT REFERENCES evidence_sources(evidence_id))"""
-    )
-    connection.executemany(
-        """INSERT INTO fetch_batches_v6(
-             batch_id,provider,connection_id,stream,request_hash,response_hash,blob_ref,
-             resource_count,budget_units,started_at,completed_at,terminal_status,evidence_id)
-           VALUES(?,?,?,?,?,?,?,?,?,?,?,?,NULL)""",
-        [
-            (
-                batch_id,
-                row["provider"],
-                row["connection_id"],
-                row["stream"],
-                row["request_hash"],
-                response_hash,
-                row["blob_ref"],
-                row["resource_count"],
-                row["budget_units"],
-                row["started_at"],
-                row["completed_at"],
-                row["terminal_status"],
-            )
-            for batch_id, (row, response_hash) in sorted(canonical.items())
-        ],
-    )
-
-
-def _rewrite_projection_aliases_v6(
-    connection: sqlite3.Connection,
-    migrated: list[tuple[sqlite3.Row, str, str]],
-) -> None:
-    aliases = [
-        (batch_id, row["batch_id"])
-        for row, batch_id, _ in migrated
-        if row["batch_id"] != batch_id
-    ]
-    for statement in (
-        "UPDATE objects SET batch_id=? WHERE batch_id=?",
-        "UPDATE activities SET batch_id=? WHERE batch_id=?",
-        "UPDATE media SET batch_id=? WHERE batch_id=?",
-        "UPDATE coverage_records SET batch_id=? WHERE batch_id=?",
-        "UPDATE tombstones SET batch_id=? WHERE batch_id=?",
-    ):
-        connection.executemany(statement, aliases)
-
-
-def _migrate_fetch_batches_v6(connection: sqlite3.Connection) -> None:
-    """Make raw-envelope hashes canonical batch IDs without body-hash uniqueness."""
-    connection.execute("DROP TRIGGER IF EXISTS fetch_batch_evidence_ai")
-    connection.execute("DROP VIEW IF EXISTS canonical_evidence_projections")
-    migrated = _canonical_fetch_rows_v6(connection)
-    _create_fetch_batches_v6(connection, _deduplicate_fetch_rows_v6(migrated))
-    connection.execute("DROP TABLE fetch_batches")
-    connection.execute("ALTER TABLE fetch_batches_v6 RENAME TO fetch_batches")
-    _rewrite_projection_aliases_v6(connection, migrated)
-
-
-def _orphan_projection_batch_ids(connection: sqlite3.Connection) -> list[str]:
-    rows = connection.execute(
-        """SELECT p.batch_id FROM (
-             SELECT batch_id FROM objects UNION SELECT batch_id FROM activities
-             UNION SELECT batch_id FROM media UNION SELECT batch_id FROM coverage_records
-             UNION SELECT batch_id FROM tombstones
-           ) p LEFT JOIN fetch_batches f ON f.batch_id=p.batch_id
-           WHERE f.batch_id IS NULL ORDER BY p.batch_id"""
-    ).fetchall()
-    return [str(row["batch_id"]) for row in rows]
-
-
-def _raw_envelope_for_batch(
-    connection: sqlite3.Connection, batch_id: str
-) -> tuple[dict[str, object], str]:
-    if re.fullmatch(r"[0-9a-f]{64}", batch_id) is None:
-        raise SocialStoreError("legacy social projection has an invalid batch ID")
-    root = _store_root(connection)
-    raw_root = root / "sources" / "social" / "raw"
-    candidates = list(raw_root.glob(f"*/*/{batch_id}.json.gz"))
-    if len(candidates) != 1:
-        raise SocialStoreError(
-            "legacy social projection raw evidence is missing or ambiguous"
-        )
-    blob_ref = candidates[0].relative_to(root).as_posix()
-    payload, digest, provider, connection_id = _read_raw_payload(
-        connection, blob_ref
-    )
-    if digest != batch_id:
-        raise SocialStoreError("legacy social raw evidence hash does not match")
-    envelope = _collector_envelope(
-        payload, provider, connection_id, required=True
-    )
-    if envelope is None:
-        raise SocialStoreError("legacy social raw evidence envelope is invalid")
-    return envelope, blob_ref
-
-
-def _recover_orphaned_fetch_batches_v6(connection: sqlite3.Connection) -> None:
-    for batch_id in _orphan_projection_batch_ids(connection):
-        envelope, blob_ref = _raw_envelope_for_batch(connection, batch_id)
-        resource_count = connection.execute(
-            """SELECT (SELECT count(*) FROM objects WHERE batch_id=?) +
-                      (SELECT count(*) FROM activities WHERE batch_id=?) +
-                      (SELECT count(*) FROM media WHERE batch_id=?)""",
-            (batch_id, batch_id, batch_id),
-        ).fetchone()[0]
-        connection.execute(
-            """INSERT INTO fetch_batches(
-                 batch_id,provider,connection_id,stream,request_hash,response_hash,
-                 blob_ref,resource_count,budget_units,started_at,completed_at,
-                 terminal_status)
-               VALUES(?,?,?,?,?,?,?,?,?,?,?,?)""",
-            (
-                batch_id,
-                envelope["provider"],
-                envelope["connection_id"],
-                envelope["stream"],
-                envelope["request_hash"],
-                envelope["response_sha256"],
-                blob_ref,
-                resource_count,
-                0,
-                envelope["observed_at"],
-                envelope["observed_at"],
-                "legacy_recovered",
-            ),
-        )
 
 
 def _migrate_source_contract(connection: sqlite3.Connection) -> None:
